add pet update endpoints
